Remove firefox config, it shouldn't be tracked
firefox configuration should probably not be tracked, don't know what
there is in it, may be something private somewhere.  Simlinking it is
hard as paths are dynamic and profiles.ini is dependant on installation
and hardening firefox became so easy and fast that version control of
its config is not necessary.